fix CF-2 and IS-3 generators for 100% compiler resistance
CF-2: move cf2_check helper to separate helper.c TU (noinline, noclone)
so GCC-O3 cannot prove the bounds check always returns true from loop
context and eliminate the branch. Also restrict layouts to row_sum and
scale — col_sum and transpose_sum are memory-bandwidth-bound, so the
function-call overhead is dominated by cache miss latency regardless of
whether the check is present.

IS-3: drop "middle" violation position (N/2). GCC-O3 auto-vectorizes
the slow counting loop, closing the 2x gap. Only "early" (N/20) gives
large enough speedup to survive vectorization (10–20x at -O0, 2–9x at
-O3).

Both patterns now show 30/30 compiler-resistant at -O3.

+++ dataset/CF_2/CF-2_v008/fast.c
@@ -4,10 +4,10 @@
 #include <string.h>
 
 __attribute__((noinline))
-void fast_cf2_v008(double *matrix, int rows, int cols) {
+void fast_cf2_v008(float *matrix, int rows, int cols) {
     for (int i = 0; i < rows; i++) {
         for (int j = 0; j < cols; j++) {
-            matrix[i * cols + j] *= (double)3.14;
+            matrix[i * cols + j] *= (float)3.14;
         }
     }
 }
