+++ dataset/IS_3/IS-3_v017/fast.c
@@ -4,7 +4,7 @@
 #include <string.h>
 
 __attribute__((noinline))
-int fast_is3_v017(double *arr, int n, double thresh) {
+int fast_is3_v017(float *arr, int n, float thresh) {
     for (int i = 0; i < n; i++) {
         if (arr[i] > thresh) return 0;
     }
