+++ dataset/IS_3/IS-3_v017/slow.c
@@ -4,7 +4,7 @@
 #include <string.h>
 
 __attribute__((noinline))
-int slow_is3_v017(double *arr, int n, double thresh) {
+int slow_is3_v017(float *arr, int n, float thresh) {
     int count = 0;
     for (int i = 0; i < n; i++) {
         if (arr[i] > thresh) count++;
